Show teams' current streaks in pre-game stats
This requires an initially higher info panel, that is shrunk when
playback starts, so that we can still fit as many games as possible on
mobile screens.

The 'reduced-height' class name needs to be added dynamically outside
Cycle.js rendering: looks like a DOM element is no longer considered the
same element if its class name changes, which results in the element being
removed and re-added, so any transitions wouldn't work.