Show team records instead of streaks and playoff spot point differences
The streaks and standings information from nhl-score-api currently returns
the info after the game, not before it (see
https://github.com/peruukki/nhl-score-api/issues/2), so showing them would
spoil how the games turn out.

Also, at the beginning of the season the records are interesting so that
you see how many games each team has played and you get a better idea of
how they are doing.