Bring games that have a goal scored into focus
This makes it a lot easier to follow all games when there are too many to
fit on the screen at the same time.

This isn't a perfect solution, as the latest goal is quickly updated before
the game is brought into focus, but changing that would require major
refactoring, maybe later.