Create more different kinds of game update events
Currently all games' latest goals are updated at the same time based on the
current clock value. Generating different kinds of game update events is
the first step in allowing us to control the playback better: we can use
the events to create per-game streams that can update sequentially in case
goals are scored during one clock tick in several games.

For now, the benefit is that this simplifies the game update animations
in score-panel.js.