Ensure correctly that final seconds are always included
If an overtime goal, that also was the final goal in any game, was scored
within the final two seconds of a minute, it was not shown because the
clock didn't advance to the last second of the final minute (due to the
clock advancing by three second steps). Improve the check for excluded
final seconds to fix this.