Keep play/pause button focus when clicked
Use snabbdom's class module to toggle play/pause button class names.
This reuses the same DOM element and maintains its focus when the button
is pressed.

Looks like the class module class names are not included in the node
selector in tests, which is just as well. I checked that they are
included in the DOM when running the app.