Upgrade devDependencies to latest
Note about `event.target as HTMLButtonElement` change to
`event.target as Node`: The original code caused the ESLint error
"Unnecessary optional chain on a non-nullish value" via the
no-unnecessary-condition rule. It seems that the ESLint rule was able to
detect that the `textContent` property can only be `null` for a document
or doctype node (https://developer.mozilla.org/en-US/docs/Web/API/Node/textContent#value),
which caused the error, but the TypeScript type still contains `null` as a
possible value, so we couldn't remove the optional chain because that would
introduce the "'button.textContent' is possibly 'null'" TypeScript error.

Casting to `Node` instead solves this because then both ESLint and
TypeScript `textContent` nullable.